Ignore non-valid short url hashes
base_convert under php7.4 gives a deprecation warning if
non-digit/letters are used.
Just keep the hash as unknown, it possibly come from fuzzing tests, but
not from a human.
This is a behaviour change, before this change the invalid signs in the
hash are ignored by base_convert and can return an existing page.

Bug: T318945
Change-Id: I0e166cf939d18798418f2252a27c678d2dd7eb6f